Read sufficient bytes in StreamReaderBuffer
The `StreamReader.read` method does not guarantee that `n` requested
bytes are returned; it could be less. (This is no very explicit in the
documentation page just saying 'up to', but the docstring is more
explicit about it. Cp. python/cpython#100226) This would result in not
reading as much bytes as would be necessary for an index or slice
requested with `StreamReaderBuffer.get()` and can cause improper parsing
when using the `parse_incremental` function.

Note that the default `n = -1` will read until EOF and not return any
less bytes. Thus, that usage two line above does not need to be fixed.